+++ tests/Protocols/Mindustry.php
@@ -18,6 +18,7 @@ namespace GameQ\Tests\Protocols;
 
 use GameQ\Exception\ProtocolException;
 use GameQ\Exception\ServerException;
+use ReflectionException;
 
 /**
  * @author Sascha Greuel <sascha@softcreatr.de>
@@ -25,7 +26,7 @@ use GameQ\Exception\ServerException;
 class Mindustry extends Base
 {
     /**
-     * @throws \ReflectionException
+     * @throws ReflectionException
      * @throws ServerException
      */
     public function testDiscoveryResponse(): void
@@ -52,7 +53,7 @@ class Mindustry extends Base
     }
 
     /**
-     * @throws \ReflectionException
+     * @throws ReflectionException
      * @throws ServerException
      */
     public function testTruncatedStringIsRejected(): void